apache-airflow-data-pipelines/requirements.txt

psycopg2-binary
pandas
numpy
jupyter
ipython-sql
keyring
boto3
botocore<1.19.0,>=1.18.0
colorama<0.4.4
docutils<0.16,>=0.10
PyYAML<5.4,>=3.10
flask-sqlalchemy

## Airflow
apache-airflow==2.0.1
apache-airflow-providers-cncf-kubernetes==1.0.2
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.0
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-postgres==1.0.1
apache-airflow-providers-sqlite==1.0.1
apache-airflow-upgrade-check==1.3.0
apache-airflow-providers-amazon==1.2.0

#psycopg2
#setuptools
#typing_extensions
#cryptography
#celery